test(*): make test database class fixture
Making the test program database a class fixture instead of a session
fixture allows grouping of tests in a class according to the type of
actions (e.g., select, insert, delete, etc.).  This makes it much
easier to have known states.

* Test databases now created using SQL file; makes it easier to add
records in the future.
* Removed function from DAO to create test database; had nothing to do
with RAMSTK in the user's hand.
* Update revision, function, and hardware controller tests to group
tests into classes of similar actions.